refactor: move ActivityEntity to Entities package
