+++ src/main/java/com/alkemy/ong/data/entities/ActivityEntity.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.data.entity;
+package com.alkemy.ong.data.entities;
 
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
