+++ src/main/java/com/alkemy/ong/data/repositories/ActivityRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.data.repository;
 
-import com.alkemy.ong.data.entity.ActivityEntity;
+import com.alkemy.ong.data.entities.ActivityEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
